Bug: Notifications caused infinite file changes
We saved the procrastination file every time after checking it for
notifications, even if it was not changed. This save would trigger a
file modification event which in turn would trigger us to check for
notifications, causing an infinte loop.

We now check if the file was modified before changing it, thereby
ensuring that we no longer just use up all available CPU time.

We still check notifications twice, because we have to save the file
after a successfull notification event to update the timestamp, but this
is a huge improvement. Improving this further would require a rework in
how we store persistant state.